requirements.txt
pandas
matplotlib
seaborn
streamlit
google-generativeai
openpyxl
